Substring layout
This patch adds an `iter_substr` method which attempts to reuse as much
of the layout as possible. As a next step, it will use the
unsafe_to_break flag to be more aggressive, but the purpose of this
commit is to flesh out the interface, and also be a plausible
implementation of the fallback case.